+++ test-support/src/main/kotlin/com/squareup/test/support/AndroidTestUtils.kt
@@ -63,3 +63,25 @@ public fun generateLibraryBuild(temporaryFolder: File): File {
   }
   return buildDirectory
 }
+
+// Helper function for generating a bare-bone Android test project
+public fun generateTestBuild(temporaryFolder: File): File {
+  val buildDirectory = temporaryFolder.canonicalFile.apply { mkdirs() }
+  File(buildDirectory, "build.gradle").apply {
+    writeText("""
+        apply plugin: 'com.android.test'
+        apply plugin: 'kotlin-android'
+
+        android {
+            targetProjectPath = ":app"
+            namespace 'com.squareup.test.library'
+            compileSdk 33
+            defaultConfig {
+                targetSdk 17
+                minSdk 17
+            }
+        }
+      """.trimIndent())
+  }
+  return buildDirectory
+}